requirements.txt
pyserial==3.2.1
RPi.GPIO==0.6.4
pynmea2==1.12.0